requirements.txt
python-telegram-bot==21.6
neonize==0.2.0
supabase==2.7.4
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.4.0
aiohttp==3.9.5
aiofiles==23.2.1
fastapi==0.115.0
uvicorn==0.30.6
<<<<<<< HEAD
slowapi==0.1.9
=======
slowapi==0.1.9
>>>>>>> 937f2086d73be9b44218523290134a49f8c47d3e
